requirements.txt
streamlit
scikit-learn
pandas
numpy
faiss-cpu
groq
python-dotenv
xgboost